Game Logic 1.0
SpacesContainer renders an isometric grid with symbols, button, and players all in-line styled to grid rows and columns. Clicking the summon button checks if any players are currently on the last board  number before drawing from a pool of symbols or ending the game. Each player selects from the pool, splicing it out. Based on the selection and each players pos, the apropirate move is calculated via imported  fn whereTo. Hops then finds every board space in between. MOVE_PLAYER actions are then dispatched to each spot. Upon landing, impact symbols are styled to the same grid cell and the player's impact score increases. Each turn, the impact symbols are counted and the leading greatest numbers are added to winning arr. To win, the array.length must be 1 and include the player1. If tied, player on last spot if moved to start, game continues.